// server/src/model/SmallItem.ts
export enum SmallItem {
    Wineskin = "wineskin",
    HalfWineskin = "half_wineskin",
    Telescope = "telescope",
    Brew = "brew",
    HalfBrew = "half_brew",
    Herb = "herb",
    BlueRunestone = "blue_runestone",
    YellowRunestone = "yellow_runestone",
    GreenRunestone = "green_runestone"
}